Remove uneeded chain_cert_file prop (#108)
The forwarder_grpc_tls_trust_chain_cert_file Forwarder property is no longer needed for production Cloud setups, since they now rely on a CA. If specified in forwarder.conf, it will cause the Forwarder to try and validate it. So, removing it here helps to avoid the error. 

See https://github.com/Graylog2/graylog-cloud/issues/1081 for an example of customer-specific Forwarder values with the cert omitted.