Avoid potential journal data loss when updating to 5.0
In Graylog 5.0.0 we removed the separate plugin packages and only have
"graylog-server" and "graylog-enterprise" packages from that point on.

If a user had "graylog-server" plus the enterprise plugins installed, they
have to install "graylog-enterprise" to update to 5.0.

That triggers an uninstall of the "graylog-server" package.

We don't want to remove the journal dir and other data in that case, so we
set remove_data to false to make sure users don't lose data.

The problem doesn't exist when users go from 5.0 to 5.1 in the future, so
we can enable the data removal when uninstalling the "graylog-server"
package for other versions.